// app-tienda/interfaces/productos/IProductos.ts
// Generated by https://quicktype.io

export interface IProducto {
    id:             string;
    modelo:         string;
    precio:         number;
    fabricante:     string;
    descripcion:    string;
    idCliente:      string;
    idCategoria:      string;
    idProveedor:      string;
    
}

// export interface IProducto: Producto[];